chore: add js-cookie dependency
This package is needed by one of the utility functions from cordova
branch. If we find it's not required once we progress further in the
merging work we can remove this dependency. For now it's just to
ensure the application can be built so that further testing and fixing
issues can commence.