fix: issue with white screen in study
* fix: fix issue with speech recognition crashing room

App is trying to stop the speech recognition even if content isn't a
speech recognition content. This commit prevents that from happening.

* fix: remove the malfunctioning recommended content request

* build: install speech recognition from master instead of tag

The tag this package was referencing doesn't exist in the repository
so install failed every time.

* fix: specify version 1.3.0 for speechrecognition

Previously the repository didn't have any version tag. But now it's
been added and the version can be specified in the package.json file.


Approved-by: Henrik Wassdahl
Approved-by: Gaël Hellé Hellé